feat: search votes error and loading status
