+++ plugins-structure/packages/common-ui/src/components/RecordForm/styles.module.css
@@ -82,7 +82,7 @@
   -moz-appearance: textfield;
 }
 
-.warning {
+.message {
   margin-top: 1rem;
   margin-bottom: 3.3rem;
 }
