+++ plugins-structure/packages/ticketvote/src/ui/TicketSearch/ModalTicketSearch.js
@@ -1,53 +1,7 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import PropTypes from "prop-types";
-import { useDispatch, useSelector } from "react-redux";
-import { Modal, Table } from "pi-ui";
-import { RecordForm } from "@politeiagui/common-ui";
-import styles from "./styles.module.css";
-import { ticketvoteResults } from "../../ticketvote/results";
-
-function TicketSearch({ placeholder, token }) {
-  const dispatch = useDispatch();
-
-  const [resultsFound, setResultsFound] = useState();
-
-  const results = useSelector((state) =>
-    ticketvoteResults.selectByToken(state, token)
-  );
-
-  function handleSearchTicket({ ticketToken }) {
-    const tickets = results
-      .filter((r) => r.ticket === ticketToken)
-      .map((ticket) => ({
-        Ticket: ticket.ticket,
-        Option: ticket.votebit === "2" ? "Yes" : "No",
-      }));
-    setResultsFound(tickets);
-  }
-
-  useEffect(() => {
-    dispatch(ticketvoteResults.fetch({ token }));
-  }, [dispatch, token]);
-
-  return (
-    <div>
-      <RecordForm onSubmit={handleSearchTicket} className={styles.searchForm}>
-        {({ TextInput }) => (
-          <TextInput
-            name="ticketToken"
-            placeholder={placeholder}
-            data-testid="ticketvote-modal-ticket-search-input"
-          />
-        )}
-      </RecordForm>
-      {resultsFound && (
-        <div data-testid="ticketvote-modal-ticket-search-table">
-          <Table headers={["Ticket", "Option"]} data={resultsFound} />
-        </div>
-      )}
-    </div>
-  );
-}
+import { Modal } from "pi-ui";
+import { TicketSearch } from "./TicketSearch";
 
 export function ModalTicketSearch({
   placeholder,
@@ -65,7 +19,7 @@ export function ModalTicketSearch({
 
 ModalTicketSearch.propTypes = {
   token: PropTypes.string.isRequired,
-  placeholder: PropTypes.node,
+  placeholder: PropTypes.string,
   onClose: PropTypes.func,
   title: PropTypes.node,
   show: PropTypes.bool,
@@ -73,5 +27,4 @@ ModalTicketSearch.propTypes = {
 
 ModalTicketSearch.defaultProps = {
   title: "Search Ticket Vote",
-  placeholder: "Search by ticket token",
 };
