+++ plugins-structure/packages/ticketvote/src/ui/TicketSearch/index.js
@@ -1,1 +1,2 @@
 export * from "./ModalTicketSearch";
+export * from "./TicketSearch";
